+++ am/serviceaccount-test.json
@@ -1,0 +1,37 @@
+{
+  "result": [
+    {
+      "_id": "123",
+      "_rev": "456",
+      "cn": [
+        "Policy Engine Service Account"
+      ],
+      "mail": [
+        "example@example.com"
+      ],
+      "username": "service_account.policy",
+      "inetUserStatus": [
+        "Active"
+      ]
+    },
+    {
+      "_id": "124",
+      "_rev": "876",
+      "cn": [
+        "IG Service Account"
+      ],
+      "mail": [
+        "ig@acme.com"
+      ],
+      "username": "service_account.ig",
+      "inetUserStatus": [
+        "Active"
+      ]
+    }
+  ],
+  "resultCount": 2,
+  "pagedResultsCookie": null,
+  "totalPagedResultsPolicy": "NONE",
+  "totalPagedResults": -1,
+  "remainingPagedResults": 0
+}
